services/deepgram-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.9.0
pydantic-settings==2.1.0
redis==4.6.0
deepgram-sdk==3.2.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
openai>=1.12.0
pyyaml>=6.0
celery[redis]==5.3.4
boto3==1.34.0

